+++ backends/gl/GlslEmit.cpp
@@ -72,7 +72,7 @@ GlslOpCodeVecs::GlslOpCodeVecs(const GlslMapping& mapping)
 
   // Dices up the code string into references #i or $i to 
   // src variables and the code fragments between references. 
-  unsigned i, j;
+  size_t i, j;
   i = j = 0;
   for (; (j = code.find_first_of("$", i)) != string::npos;) {
     frag.push_back(code.substr(i, j - i));
